changing frontend design of logging in
