+++ frontend/src/App.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Routes, Route, useLocation } from 'react-router-dom';
 import { Helmet, HelmetProvider } from 'react-helmet-async';
 import { CssBaseline, Box, ThemeProvider } from '@mui/material';
+import InitalPage from './components/InitalPage';
 import Login from './components/Login';
 import OAuthCallbackHandler from './components/OAuthCallbackHandler';
 import Dashboard from './components/Dashboard';
@@ -16,7 +17,7 @@ import OrdersToCheckPage from './components/OrdersToCheckPage';
 
 const App: React.FC = () => {
   const location = useLocation();
-  const isLoginPage = location.pathname === '/';
+  const isInitalPage = location.pathname === '/';
 
   return (
     <ThemeProvider theme={theme}>
@@ -28,10 +29,11 @@ const App: React.FC = () => {
           </Helmet>
           <CssBaseline />
           <Box sx={{ display: 'flex', flexDirection: 'column', minHeight: '100vh', width: '100%' }}>
-            <TopBar isLoginPage={isLoginPage} />
+            <TopBar isInitalPage={isInitalPage} />
             <Box sx={{ flexGrow: 1, width: '100%' }}>
               <Routes>
-                <Route path="/" element={<Login />} />
+                <Route path="/" element={<InitalPage />} />
+                <Route path="/login" element={<Login />} />
                 <Route path="/oauth/callback" element={<OAuthCallbackHandler />} />
                 <Route path="/dashboard" element={<Dashboard />} />
                 <Route path="/dashboard/settings" element={<Settings />} />
